Rename project from playlist-creator to apple-music-mcp
Broadens scope from playlist-only CLI to general Apple Music MCP server.

- Rename package: playlist_creator/ → apple_music_mcp/
- Rename MCP entry point: playlist-creator-mcp → apple-music-mcp
- Keep playlist-creator CLI entry point intact
- Update all imports, docs, and tests
