+++ apple_music_mcp/__init__.py
@@ -1,0 +1,3 @@
+"""Apple Music MCP server and CLI — search catalog, manage playlists, and more."""
+
+__version__ = "0.1.0"
