+++ apple_music_mcp/auth.py
