--- playlist_creator/__init__.py
@@ -1,3 +1,0 @@
-"""Create Apple Music playlists from markdown files."""
-
-__version__ = "0.1.0"
